IS-1978: Move /sykmeldinger into /sider folder
