+++ src/components/speiling/sykepengsoknader/soknad-selvstendig/SendtSoknadSelvstendigStatuspanel.tsx
@@ -5,7 +5,7 @@ import Statuspanel, {
 } from "../../Statuspanel";
 import { VerktoyKnapp, Verktoylinje } from "../../Verktoylinje";
 import { tilLesbarDatoMedArstall } from "@/utils/datoUtils";
-import SykmeldingNokkelOpplysning from "../../sykmeldinger/sykmelding/sykmeldingOpplysninger/SykmeldingNokkelOpplysning";
+import SykmeldingNokkelOpplysning from "@/sider/sykmeldinger/sykmelding/sykmeldingOpplysninger/SykmeldingNokkelOpplysning";
 import { SykepengesoknadDTO } from "@/data/sykepengesoknad/types/SykepengesoknadDTO";
 import { erOpprettetSisteAar } from "@/utils/sykepengesoknadUtils";
 import SykepengerOgSaksbehandlingstiderLink from "@/utils/soknad-felles/SykepengerOgSaksbehandlingstiderLink";
