+++ src/routers/AppRouter.tsx
@@ -1,9 +1,9 @@
 import React, { ReactElement } from "react";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import AktivBrukerTilgangLaster from "@/components/AktivBrukerTilgangLaster";
-import SykmeldingerContainer from "../components/speiling/sykmeldinger/container/SykmeldingerContainer";
+import SykmeldingerContainer from "@/sider/sykmeldinger/container/SykmeldingerContainer";
 import SykepengesoknaderContainer from "../components/speiling/sykepengsoknader/container/SykepengesoknaderSide";
-import DinSykmeldingContainer from "../components/speiling/sykmeldinger/container/DinSykmeldingContainer";
+import DinSykmeldingContainer from "@/sider/sykmeldinger/container/DinSykmeldingContainer";
 import HistorikkContainer from "@/sider/historikk/container/HistorikkContainer";
 import { erGyldigFodselsnummer } from "@/utils/frnValideringUtils";
 import MotelandingssideContainer from "@/sider/mote/container/MotelandingssideContainer";
