+++ src/sider/sykmeldinger/sykmelding/DinSykmelding.tsx
@@ -3,11 +3,11 @@ import { Link } from "react-router-dom";
 import Alertstripe from "nav-frontend-alertstriper";
 import { SykmeldingOldFormat } from "@/data/sykmelding/types/SykmeldingOldFormat";
 import DineSykmeldingOpplysninger from "./sykmeldingOpplysninger/DineSykmeldingOpplysninger";
-import Hjelpeboble from "../../../bjorn/Hjelpeboble";
+import Hjelpeboble from "../../../components/bjorn/Hjelpeboble";
 import {
   PersonHighContrastImage,
   PersonImage,
-} from "../../../../../img/ImageComponents";
+} from "../../../../img/ImageComponents";
 
 const texts = {
   eldreSykmeldinger: "Du har eldre sykmeldinger som du bør behandle før denne.",
