+++ src/sider/sykmeldinger/sykmelding/koronasykmeldinger/KoronaSykmelding-Avbrutt.tsx
