+++ src/sider/sykmeldinger/sykmelding/koronasykmeldinger/KoronaSykmelding-Bekreftet.tsx
