+++ src/sider/sykmeldinger/sykmelding/koronasykmeldinger/KoronaSykmelding-Ny.tsx
@@ -3,8 +3,8 @@ import { Undertittel, Normaltekst } from "nav-frontend-typografi";
 import { Knapp } from "nav-frontend-knapper";
 import { SykmeldingOldFormat } from "@/data/sykmelding/types/SykmeldingOldFormat";
 import DineKoronaSykmeldingOpplysninger from "../sykmeldingOpplysninger/DineKoronaSykmeldingOpplysninger";
-import Hjelpeboble from "../../../../bjorn/Hjelpeboble";
-import { PersonImage } from "../../../../../../img/ImageComponents";
+import Hjelpeboble from "../../../../components/bjorn/Hjelpeboble";
+import { PersonImage } from "../../../../../img/ImageComponents";
 
 const texts = {
   pageSubtitle: "for selvstendig næringsdrivende og frilansere",
