+++ src/sider/sykmeldinger/sykmelding/sykmeldingOpplysninger/SykmeldingCheckboxForFelt.tsx
