+++ src/sider/sykmeldinger/sykmelding/sykmeldingOpplysninger/SykmeldingPeriode.tsx
