+++ src/sider/sykmeldinger/sykmelding/sykmeldingOpplysninger/SykmeldingPerioder.tsx
