+++ src/sider/sykmeldinger/sykmelding/sykmeldingOpplysninger/flereopplysninger/BedreArbeidsevne.tsx
