+++ src/sider/sykmeldinger/sykmeldinger/SykmeldingTeasere.tsx
